add support of additional data of anonymous resource collection
