+++ src/Support/TypeToSchemaExtensions/JsonResourceTypeToSchema.php
@@ -28,6 +28,9 @@ use Illuminate\Http\Resources\MissingValue;
 
 class JsonResourceTypeToSchema extends TypeToSchemaExtension
 {
+    use FlattensMergeValues;
+    use MergesOpenApiObjects;
+
     public function shouldHandle(Type $type)
     {
         return $type instanceof ObjectType
@@ -67,102 +70,6 @@ class JsonResourceTypeToSchema extends TypeToSchemaExtension
         $array->isList = KeyedArrayType::checkIsList($array->items);
 
         return $this->openApiTransformer->transform($array);
-    }
-
-    private function flattenMergeValues(array $items)
-    {
-        return collect($items)
-            ->flatMap(function (ArrayItemType_ $item) {
-                if ($item->value instanceof KeyedArrayType) {
-                    $item->value->items = $this->flattenMergeValues($item->value->items);
-                    $item->value->isList = KeyedArrayType::checkIsList($item->value->items);
-
-                    return [$item];
-                }
-
-                if (
-                    $item->value instanceof Union
-                    && (new TypeWalker)->first($item->value, fn (Type $t) => $t->isInstanceOf(Carbon::class))
-                ) {
-                    (new TypeWalker)->replace($item->value, function (Type $t) {
-                        return $t->isInstanceOf(Carbon::class)
-                            ? tap(new StringType, fn ($t) => $t->setAttribute('format', 'date-time'))
-                            : null;
-                    });
-
-                    return [$item];
-                }
-
-                if ($item->value->isInstanceOf(JsonResource::class)) {
-                    $resource = $this->getResourceType($item->value);
-
-                    if ($resource->isInstanceOf(MissingValue::class)) {
-                        return [];
-                    }
-
-                    if (
-                        $resource instanceof Union
-                        && (new TypeWalker)->first($resource, fn (Type $t) => $t->isInstanceOf(MissingValue::class))
-                    ) {
-                        $item->isOptional = true;
-
-                        return [$item];
-                    }
-                }
-
-                if (
-                    $item->value instanceof Union
-                    && (new TypeWalker)->first($item->value, fn (Type $t) => $t->isInstanceOf(MissingValue::class))
-                ) {
-                    $newType = array_filter($item->value->types, fn (Type $t) => ! $t->isInstanceOf(MissingValue::class));
-
-                    if (! count($newType)) {
-                        return [];
-                    }
-
-                    $item->isOptional = true;
-
-                    if (count($newType) === 1) {
-                        $item->value = $newType[0];
-
-                        return $this->flattenMergeValues([$item]);
-                    }
-
-                    $item->value = new Union($newType);
-
-                    return $this->flattenMergeValues([$item]);
-                }
-
-                if (
-                    $item->value instanceof Generic
-                    && $item->value->isInstanceOf(MergeValue::class)
-                ) {
-                    $arrayToMerge = $item->value->templateTypes[1];
-
-                    // Second generic argument of the `MergeValue` class must be a keyed array.
-                    // Otherwise, we ignore it from the resulting array.
-                    if (! $arrayToMerge instanceof KeyedArrayType) {
-                        return [];
-                    }
-
-                    $arrayToMergeItems = $this->flattenMergeValues($arrayToMerge->items);
-
-                    $mergingArrayValuesShouldBeRequired = $item->value->templateTypes[0] instanceof LiteralBooleanType
-                        && $item->value->templateTypes[0]->value === true;
-
-                    if (! $mergingArrayValuesShouldBeRequired || $item->isOptional) {
-                        foreach ($arrayToMergeItems as $mergingItem) {
-                            $mergingItem->isOptional = true;
-                        }
-                    }
-
-                    return $arrayToMergeItems;
-                }
-
-                return [$item];
-            })
-            ->values()
-            ->all();
     }
 
     /**
@@ -215,25 +122,12 @@ class JsonResourceTypeToSchema extends TypeToSchemaExtension
         return new Reference('schemas', $type->name, $this->components);
 
         /*
-         * @todo: Allow (enforce) user to explicitly pass short and unique names for the reference.
+         * @todo: Allow (enforce) user to explicitly pass short and unique names for the reference and avoid passing components.
          * Otherwise, only class names are correctly handled for now.
          */
         return Reference::in('schemas')
             ->shortName(class_basename($type->name))
             ->uniqueName($type->name);
-    }
-
-    private function mergeOpenApiObjects(OpenApiTypes\ObjectType $into, OpenApiTypes\Type $what)
-    {
-        if (! $what instanceof OpenApiTypes\ObjectType) {
-            return;
-        }
-
-        foreach ($what->properties as $name => $property) {
-            $into->addProperty($name, $property);
-        }
-
-        $into->addRequired(array_keys($what->properties));
     }
 
     private function getResourceType(Type $type): Type
